requirements.txt
faiss-cpu==1.9.0
huggingface-hub==0.36.0
keybert==0.9.0
newsapi-python==0.2.7
numpy==1.26.4
ollama==0.6.1
pandas==2.3.3
python-dotenv==1.1.0
PyYAML==6.0.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
sentence-transformers==5.2.0
streamlit==1.54.0
tokenizers==0.22.1
torch==2.6.0+cu126
tqdm==4.67.1
transformers==4.57.3